requirements.txt
absl-py==0.9.0
astunparse==1.6.3
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1592842333140/work
decorator==4.4.2
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py @ file:///tmp/build/80754af9/h5py_1593454121459/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.8.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846044287/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.4
nibabel==2.5.1
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==7.1.2
protobuf==3.12.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy @ file:///tmp/build/80754af9/scipy_1592930540905/work
six==1.13.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
